Adjust comparing User and Containers to care about backends too
Do not treat different versions of the same User or Container objects
as equal. Compare manifests catalog (for User) and backends (for
Container). This avoids the results deduplication to remove actually
different containers from the results.

This is only half-solution, because ideally we'd like to detect which
container manifest instance is the "new" one. But currently the manifest
content has no information to make such decision, so - let the code to
return all of them.

Related to #461