search: do not re-load local users to the client
Use already loaded users into the Client instance. Since
add3dc500e5cc1618cbab18b9d642fdee5d27489, they are loaded in __init__,
unless explicitly requested to not do so - respect this request in
search too.

Specifically, when the user is loaded via a specific bridge (in
Search._bridge_step), use that object. It may be loaded from a different
version of the User manifest than the one stored locally. This is
especially important when resolving 'reference-container' in a storage -
use the same User manifest as the one used to load the parent container.

Fix #461