link: distinguish link owner and target object owner
When loading an object via a link object, user context of the link
(where the link appears, like - user manifest) for handling the storage,
instead of the user expected as the target object. This fixes loading
user manifest from a bridge: the link object is owned by one user, while
the target object (user manifest) is owned by another.

This makes the load_object_from_dict() function take both 'owner'
(context of the object in 'dictionary') and 'expected_owner' (owner of
the thing to load). Outside of the bridge usage, they will be the same.